refactor: 補完ロジックをWorkEnrichmentServiceに分離しlimit対応を追加
